Updated file permissions (Android 13)
Updating Android SDK version to 34, with a minimum SDK version increased to 30.

Modified the AndroidManifest.xml to grant the READ_EXTERNAL_STORAGE and MANAGE_EXTERNAL_STORAGE permissions

App now performs a check on boot to check if the ACTION_MANAGE_APP_ALL_FILES_ACCESS_PERMISSION is already granted, if not, will prompt android to grant this for the URI of this package.

Handling of the temporary file has now changed. Due to changes in Android 13, the permissions to be able to access /data/user/... are not the same. Permissions now allow the temporary file to be written to the folder where the desired ROM save location is. The temporary ROM will also be copied to the desired ROM, and then deleted.